chore(refactor): relocate directories appropriately
