+++ package.json
@@ -4,15 +4,15 @@
   "description": "full stack web messaging application",
   "scripts": {
     "babel-version": "babel --version",
-    "remove-dist": "node_modules/.bin/rimraf ./client/dist",
+    "build": "npm run remove-dist && webpack",
     "coveralls": "nyc --reporter=lcov --reporter=text-lcov npm test",
-    "exec": "nodemon ./server/bin/www --exec babel-node app",
-    "build": "npm run remove-dist && webpack -p",
-    "start": "babel-node ./server/bin/www --exec babel-node app",
+    "exec": "babel-node ./server/bin/www --exec babel-node app",
+    "remove-dist": "node_modules/.bin/rimraf ./client/dist",
+    "start": "nodemon ./server/bin/www --exec babel-node ./server/app",
     "test_": "nyc mocha --compilers js:babel-core/register --timeout 10000",
     "test-dev": "mocha --compilers js:babel-core/register --timeout 10000",
     "test": "nyc --reporter=html --reporter=text mocha --compilers js:babel-register",
-    "start-dev": "webpack-dev-server && webpack-dev-server"
+    "start-dev": "webpack && webpack-dev-server"
   },
   "repository": {
     "type": "git",
@@ -44,7 +44,9 @@
     "mocha-lcov-reporter": "^1.3.0",
     "nyc": "^11.0.3",
     "webpack": "^3.4.1",
-    "webpack-dev-server": "^2.6.1"
+    "webpack-dev-middleware": "^1.12.0",
+    "webpack-dev-server": "^2.6.1",
+    "webpack-hot-middleware": "^2.18.2"
   },
   "dependencies": {
     "autoprefixer": "7.1.1",
@@ -117,10 +119,7 @@
     "style-loader": "^0.18.2",
     "sw-precache-webpack-plugin": "0.11.3",
     "url-loader": "0.5.9",
-    "webpack": "^3.4.1",
-    "webpack-dev-middleware": "^1.12.0",
-    "webpack-dev-server": "^2.6.1",
-    "webpack-hot-middleware": "^2.18.2"
+    "webpack": "^3.4.1"
   },
   "jest": {
     "collectCoverageFrom": [
